refactor: replaced commonJS imports to ESM in esm.msj
